Remove many uses of the new keyword
The new keyword creates unprotected raw pointers with the assumption
that the caller takes ownership of the returned value. That's not
great, and std::make_unique is a much more elegant solution. Therefore,
we take as many uses of the new keyword as possible and replace them by
std::make_unique.