Remove various uses of raw pointers (part 1)
After previous refactoring, the code still contains many uses of the
new and delete keywords, as well as many cases where raw pointers are
used to express ownership. To improve the code, we get rid of these
using smart pointers (in most cases, std::unique_ptr). The goal of
these changes is to reduce code volume, increase clarity when it comes
to memory ownership, improve interface usability, and improve memory
safety. Since there are many remaining use cases, we split this commit
up into many parts. This is to keep the individual commits relatively
small and to allow easier bisecting and debugging if a bug should be
found in these diffs.