Use std::unique_ptr for GXFTrajectory internals
Previously, the internal representation of track states in the
GXFTrajectory class was a vector of raw pointers. We replace this by a
vector of smart pointers, making the ownership of these states more
explicit, and simplifying object cleanup. This obviously involves
updating any call sites of the state accessor methods, too.